Remove duplicated delegatedMock
