+++ core/tst/software/aws/toolkits/core/utils/DelegateSdkConsumers.kt
@@ -3,13 +3,14 @@
 
 package software.aws.toolkits.core.utils
 
+import com.nhaarman.mockitokotlin2.KStubbing
+import com.nhaarman.mockitokotlin2.withSettings
 import org.mockito.Mockito
 import org.mockito.invocation.InvocationOnMock
 import org.mockito.stubbing.Answer
 import software.amazon.awssdk.core.SdkRequest
 import kotlin.reflect.full.isSubclassOf
 
-// TODO: Copy pasted from jetbrains-core, better way to share this?
 class DelegateSdkConsumers : Answer<Any> {
     override fun answer(invocation: InvocationOnMock): Any? {
         val method = invocation.method
@@ -21,4 +22,15 @@ class DelegateSdkConsumers : Answer<Any> {
             Mockito.RETURNS_DEFAULTS.answer(invocation)
         }
     }
+}
+
+inline fun <reified T : Any> delegateMock(): T = Mockito.mock(
+    T::class.java,
+    withSettings(
+        defaultAnswer = DelegateSdkConsumers()
+    )
+)
+
+inline fun <reified T : Any> delegateMock(stubbing: KStubbing<T>.(T) -> Unit): T = delegateMock<T>().apply {
+    KStubbing(this).stubbing(this)
 }
