+++ jetbrains-core/tst/software/aws/toolkits/jetbrains/services/lambda/upload/LambdaCreatorTest.kt
@@ -27,10 +27,10 @@ import software.amazon.awssdk.services.lambda.model.UpdateFunctionConfigurationR
 import software.amazon.awssdk.services.s3.S3Client
 import software.amazon.awssdk.services.s3.model.PutObjectRequest
 import software.amazon.awssdk.services.s3.model.PutObjectResponse
+import software.aws.toolkits.core.utils.delegateMock
 import software.aws.toolkits.jetbrains.core.MockClientManagerRule
 import software.aws.toolkits.jetbrains.services.iam.IamRole
 import software.aws.toolkits.jetbrains.services.lambda.LambdaBuilder
-import software.aws.toolkits.jetbrains.utils.delegateMock
 import software.aws.toolkits.jetbrains.utils.rules.JavaCodeInsightTestFixtureRule
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
