+++ jetbrains-core/tst/software/aws/toolkits/jetbrains/services/s3/DeleteBucketTestMock.kt
@@ -19,9 +19,9 @@ import software.amazon.awssdk.services.s3.model.ListObjectVersionsResponse
 import software.amazon.awssdk.services.s3.model.ObjectVersion
 import software.amazon.awssdk.services.s3.paginators.ListObjectVersionsIterable
 import software.aws.toolkits.core.s3.deleteBucketAndContents
+import software.aws.toolkits.core.utils.delegateMock
 import software.aws.toolkits.jetbrains.core.MockClientManagerRule
 import software.aws.toolkits.jetbrains.services.s3.bucketActions.DeleteBucketAction
-import software.aws.toolkits.jetbrains.utils.delegateMock
 import java.util.function.Consumer
 
 class DeleteBucketTestMock {
